requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
Flask==1.0.2
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
googlemaps==3.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python==8.0.11
numpy==1.15.0
pandas==0.23.3
protobuf==3.6.0
pyowm==2.8.0
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
requests==2.18.4
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.22
Werkzeug==0.14.1
PyMySQL==0.9.2
